pyproject.toml
[project]
name = "neuro-tech"
version = "0.1.0"
description = "IPTU Data Pipeline - Neuro Tech Challenge"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.1.0",
    "pandas>=2.3.3",
    "pyarrow>=14.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "pillow>=10.0.0",
    "apache-airflow>=2.7.0",
    "plotly>=5.18.0",
    "kaleido>=0.2.1",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pydeequ>=1.5.0",
    "delta-spark>=3.3.2",
]

[project.optional-dependencies]
pyspark = [
    "pyspark>=3.4.0,<4.0.0",
    "delta-spark>=3.0.0",
    "pydeequ>=1.5.0",
]
